Created new projects for the samples to better isolate the code
