+++ snippets/csharp/VS_Snippets_Misc/tpl_partitioners/cs/partitioners.csproj
@@ -1,0 +1,13 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <Compile Remove="*.cs" />
+    <Compile Include="partitioners.cs" />
+  </ItemGroup>
+
+</Project>
